# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PayD Backend API",
    "version": "1.0.0",
    "description": "API Documentation for PayD Backend"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/scaling/query-stats": {
      "get": {
        "summary": "Recent slow-query statistics (admin only)",
        "tags": [
          "Scaling"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "minMs",
            "schema": {
              "type": "integer",
              "default": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of recent slow queries"
          }
        }
      }
    },
...
    "/api/v1/scaling/latency-percentiles": {
      "get": {
        "summary": "p50 / p95 / p99 latency per endpoint (Part 25)",
        "description": "Computes approximate latency percentiles from the api_latency_histogram table. Returns the top endpoints by p99 so you can spot regressions.\n",
        "tags": [
          "Scaling"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "windowMinutes",
            "schema": {
              "type": "integer",
              "default": 60
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Latency percentile breakdown per endpoint"
          }
        }
      }
    },
    "/api/v1/scaling/pool-history": {
      "get": {
        "summary": "Recent connection-pool health snapshots (Part 25)",
        "description": "Returns time-series snapshots from db_pool_health so operators can spot connection-pool saturation trends without needing Grafana.\n",
        "tags": [
          "Scaling"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pool health snapshots, most recent first"
          }
        }
      }
    },
...
  }
}